requirements.txt
feedparser
openai
tweepy
python-dotenv
requests
pytz
